LayoutTests/js/slash-lineterminator-parse-expected.txt

This test checks for correct handling ofr backslash-newline in JS strings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mystring is "hellothere"
PASS successfullyParsed is true

TEST COMPLETE
